Remove discord code from tests, switch to longs, remove unchecked get

+++ src/main/java/logic/DamageLogic.java
@@ -33,24 +33,24 @@ public class DamageLogic implements VelenSlashEvent {
                 .addChoice("wound", "wound")
                 .setRequired(true));
         options.add(new SlashCommandOptionBuilder()
-                .setType(SlashCommandOptionType.INTEGER)
+                .setType(SlashCommandOptionType.LONG)
                 .setName("count")
                 .setDescription("the amount of damage that is being inflicted")
                 .setRequired(true));
         options.add(new SlashCommandOptionBuilder()
-                .setType(SlashCommandOptionType.INTEGER)
+                .setType(SlashCommandOptionType.LONG)
                 .setName("stun-armor")
                 .setDescription("the amount of stun armor you have"));
         options.add(new SlashCommandOptionBuilder()
-                .setType(SlashCommandOptionType.INTEGER)
+                .setType(SlashCommandOptionType.LONG)
                 .setName("basic-armor")
                 .setDescription("the amount of basic armor you have"));
         options.add(new SlashCommandOptionBuilder()
-                .setType(SlashCommandOptionType.INTEGER)
+                .setType(SlashCommandOptionType.LONG)
                 .setName("wound-armor")
                 .setDescription("the amount of wound armor you have"));
         options.add(new SlashCommandOptionBuilder()
-                .setType(SlashCommandOptionType.INTEGER)
+                .setType(SlashCommandOptionType.LONG)
                 .setName("resilience")
                 .setDescription("the amount of resilience you have"));
